@@ -12,6 +12,7 @@
   "dependencies": {
     "@astrojs/check": "^0.9.4",
     "@astrojs/mdx": "^3.1.7",
+    "@astrojs/node": "^8.3.4",
     "@astrojs/rss": "^4.0.7",
     "@astrojs/sitemap": "^3.2.0",
     "@astrojs/svelte": "^5.7.1",
@@ -27,10 +28,10 @@
     "rehype-sanitize": "^6.0.0",
     "remark-html": "^16.0.1",
     "remark-math": "^6.0.0",
-    "svelte": "^4.2.12",
+    "svelte": "^5.0.0",
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.4.2",
+    "typescript": "^5.5.0",
     "unist-util-visit": "^5.0.0"
   },
   "devDependencies": {
